12 dots drawing

+++ SerialCommunication/SerialCommunication/Structures/Dot.cs
@@ -3,7 +3,7 @@ namespace SerialCommunication.Structures
 {
     public class Dot
     {
-        public int AngleInDegrees { get; set; }
+        public double AngleInDegrees { get; set; }
         public double Distance { get; set; }
     }
 }
